backend/requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
python-multipart
python-jose[cryptography]
passlib[bcrypt]
scikit-learn
numpy
pandas
websockets
aiofiles
httpx
python-dotenv